svn48
+world render separated from rr.cpp to render.cpp
+MATRIX operations moved from rrengine.lib to world.lib

+++ src/World/LoadWorldR3d.cpp
@@ -149,7 +149,7 @@ extern WORLD *load_world(char *name)
  if(!name) return NULL;
  f=fopen(name,"rb");
  if (!f) return NULL;
- fseek(f,strlen(sig),SEEK_SET);
+ fseek(f,(long)strlen(sig),SEEK_SET);
 
  world=ALLOC(WORLD);
